started taking the loop apart (ServerConfigServerParser.cpp:257)
