+++ config/test.conf
@@ -17,11 +17,11 @@ server {
         
 
     route "/" {
-        #auto_index
-        #enable_cgi
+        #auto_index  
+        enable_cgi
         #methods ["GET","POST","PATCH","PUT"]
         root "html"
         redirect "http://42.42.42.42"
-        index "index.html"
+        index "index.html" 
 }
 }
