+++ src/ServerConfigServerParser.cpp
@@ -226,108 +226,6 @@ void	ServerConfig::parseServerRoot(std::string curr_line, ServerCfg &server_conf
 	if (!(ltoken.empty()) && ltoken[0] != '#') throw std::runtime_error("Error: unexpected token: line: " + ParserUtils::intToString(_bad_line)); 
 }
 
-/*	@parseServerRoute:
- *		Looks for opening subkeywd_bracket ({)
- *		Calls getline parsing each subkeywd to the RouteCfg object
- *
- * */
-void	ServerConfig::parseServerRoute(std::string curr_line, ServerCfg &server_conf) {
-	RouteCfg	route_conf;
-
-	std::istringstream	iss_curr_line(curr_line);
-	std::string		token, ltoken;
-
-	std::getline(iss_curr_line, token, ' ');
-	std::getline(iss_curr_line, token, ' ');
-
-	if (token[0] != '"' || token[token.size() - 1] != '"' || token.length() < 3 || token[1] != '/')
-		throw std::runtime_error("Error: invalid route_path: line: " + ParserUtils::intToString(_bad_line));
-
-	route_conf.route_path = token;
-
-	std::getline(iss_curr_line, ltoken, ' ');
-	if (!(ltoken.empty()) && ltoken[0] != '#' && ltoken.compare("{") != 0) {
-		throw std::runtime_error("Error: unexpected token: line: " + ParserUtils::intToString(_bad_line));
-	} else if (ltoken.compare("{") == 0) _subkeywd_bracket = true;
-
-	while (!_subkeywd_bracket) {
-		std::getline(_fd_conf, curr_line); _bad_line++;
-		curr_line = ParserUtils::parseLine(curr_line, "	", " ");
-		std::istringstream	iss_curr_line(curr_line);
-		std::getline(iss_curr_line, token, ' ');
-
-		if (token.compare("{") != 0 && token[0] != '#') throw std::runtime_error("Error: missing opening bracket: line " + ParserUtils::intToString(_bad_line));
-		else if (token.compare("{") == 0) _subkeywd_bracket = true;
-	}
-
-	// maybe take this loop apart into diff funcs :/
-	while (getline(_fd_conf, curr_line)) {
-		curr_line = ParserUtils::parseLine(curr_line, "	", " ");
-		_bad_line++;
-		if (curr_line.empty()) continue ;
-
-		std::istringstream	iss_c_line(curr_line);
-		std::string		ntoken, nltoken;
-
-		std::getline(iss_c_line, ntoken, ' ');
-		if (ntoken.compare("}") == 0) {
-			_subkeywd_bracket = false; server_conf.routes.push_back(route_conf); return ;
-		} else if (ntoken.compare("auto_index") == 0) {
-			if (route_conf.auto_index == true) throw std::runtime_error("Error: multiple index definitions: line: " + ParserUtils::intToString(_bad_line));
-			route_conf.auto_index = true;
-			std::getline(iss_c_line, nltoken, ' ');
-			if (!(nltoken.empty()) && nltoken[0] != '#') throw std::runtime_error("Error: unexpected token: line: " + ParserUtils::intToString(_bad_line));
-		} else if (ntoken.compare("index") == 0) {
-			if (route_conf.index.compare("notgiven") != 0) throw std::runtime_error("Error: multiple index definitions: line: " + ParserUtils::intToString(_bad_line));
-			std::getline(iss_c_line, ntoken, ' ');
-			if (ntoken[0] != '"' || ntoken[ntoken.length() - 1] != '"' || ntoken.length() < 3) throw std::runtime_error("Error: invalid index: line: " + ParserUtils::intToString(_bad_line));
-			ntoken = ParserUtils::removeDelimiters(ntoken);
-			route_conf.index = ntoken;
-			std::getline(iss_c_line, nltoken, ' ');
-			if (!(nltoken.empty()) && nltoken[0] != '#') throw std::runtime_error("Error: unexpected token: line: " + ParserUtils::intToString(_bad_line));
-		} else if (ntoken.compare("enable_cgi") == 0) {
-			if (route_conf.cgi_enabled == true) throw std::runtime_error("Error: multiple enable_cgi definitions: line: " + ParserUtils::intToString(_bad_line));
-			route_conf.cgi_enabled = true;
-			std::getline(iss_c_line, nltoken, ' ');
-			if (!(nltoken.empty()) && nltoken[0] != '#') throw std::runtime_error("Error: unexpected token: line: " + ParserUtils::intToString(_bad_line));
-		} else if (ntoken.compare("root") == 0) {
-			if (route_conf.root.compare("nopath") != 0) throw std::runtime_error("Error: multiple root definitions: line: " + ParserUtils::intToString(_bad_line));
-			std::getline(iss_c_line, ntoken, ' ');
-			if (ntoken[0] != '"' || ntoken[ntoken.length() - 1] != '"' || ntoken.length() < 3) throw std::runtime_error("Error: invalid root path: line: " + ParserUtils::intToString(_bad_line));
-			ntoken = ParserUtils::removeDelimiters(ntoken);
-			for (int i = 0; ntoken[i] != '\0'; i++) {
-				if (ntoken[i] == '/' && ntoken[i + 1] == '/') throw std::runtime_error("Error: invalid root_path: line: " + ParserUtils::intToString(_bad_line));
-			}
-			route_conf.root = ntoken;
-			std::getline(iss_c_line, nltoken, ' ');
-			if (!(nltoken.empty()) && nltoken[0] != '#') throw std::runtime_error("Error: unexpected token: line: " + ParserUtils::intToString(_bad_line));
-		} else if (ntoken.compare("methods") == 0) {
-			if (!(route_conf.accepted_methods.empty())) throw std::runtime_error("Error: multiple methods definitions: line: " + ParserUtils::intToString(_bad_line));
-			std::getline(iss_c_line, ntoken, ' ');
-			if (ntoken[0] != '[' || ntoken[ntoken.length() - 1] != ']' || ntoken.length() < 5) throw std::runtime_error("Error: invalid methods: line: " + ParserUtils::intToString(_bad_line));
-			try { getParams(ntoken, route_conf.accepted_methods);
-			} catch (std::exception &ex) {
-				throw ;
-			}
-			std::getline(iss_c_line, nltoken, ' ');
-			if (!(nltoken.empty()) && nltoken[0] != '#') throw std::runtime_error("Error: unexpected token: line: " + ParserUtils::intToString(_bad_line));
-		} else if (ntoken.compare("redirect") == 0) {
-			if (route_conf.is_redirect == true) throw std::runtime_error("Error: multiple redirection definitions: line: " + ParserUtils::intToString(_bad_line));
-			std::getline(iss_c_line, ntoken, ' ');
-			if (ntoken[0] != '"' || ntoken[ntoken.length() - 1] != '"' || ntoken.length() < 3) throw std::runtime_error("Error: invalid redirection: line: " + ParserUtils::intToString(_bad_line));
-			ntoken = ParserUtils::removeDelimiters(ntoken);
-			if (ParserUtils::isValidURL(ntoken)) { route_conf.is_redirect = true; route_conf.redirect_to = ntoken; }
-			else
-				throw std::runtime_error("Error: invalid redirection: line: " + ParserUtils::intToString(_bad_line));
-			std::getline(iss_c_line, nltoken, ' ');
-			if (!(nltoken.empty()) && nltoken[0] != '#') throw std::runtime_error("Error: unexpected token: line: " + ParserUtils::intToString(_bad_line));
-			
-		}
-
-	}
-	if (_fd_conf.eof()) throw std::runtime_error("Error: missing closing bracket in route_config");	
-}
-
 /*	@parseServer:
  * 		Checks if inital '{' was found, if not finds it or throws error
  * 		If one subkeyword (block) is found, calls for its parsing
